nautilus_trader_service/requirements.txt
# Nautilus Trader Dependencies
nautilus-trader>=1.190.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# MetaTrader5 Integration
MetaTrader5>=5.0.45
pytz>=2023.3

# Data & Analysis
ta>=0.10.2
scipy>=1.10.0
scikit-learn>=1.3.0

# API & Communication
fastapi>=0.103.0
uvicorn>=0.23.0
websockets>=11.0
aiohttp>=3.8.0

# Database
asyncpg>=0.28.0
redis>=5.0.0

# Monitoring & Logging
prometheus-client>=0.17.0
structlog>=23.1.0